Improve historic benchmark execution.
