+++ libraries/ag-charts-test/package.json
@@ -3,6 +3,8 @@
   "version": "11.3.1-beta.20250611",
   "dependencies": {
     "skia-canvas": "^2.0.2",
+    "pixelmatch": "^5.2.1",
+    "pngjs": "3.4.0",
     "tslib": "^2.3.0"
   },
   "type": "commonjs",
