+++ libraries/ag-charts-test/src/canvas/to-match-image.ts
@@ -1,0 +1,31 @@
+import * as fs from 'fs';
+import pixelmatch from 'pixelmatch';
+import { PNG } from 'pngjs';
+
+export const CANVAS_WIDTH = 800;
+export const CANVAS_HEIGHT = 600;
+
+export function toMatchImage(this: any, actual: ImageData, expected: ImageData, { writeDiff = true } = {}) {
+    // Grab values from enclosing Jest scope.
+    const { testPath, currentTestName } = this;
+
+    const width = CANVAS_WIDTH;
+    const height = CANVAS_HEIGHT;
+    const diff = new PNG({ width, height });
+    const result = pixelmatch(actual.data, expected.data, diff.data, width, height, { threshold: 0.01 });
+
+    const diffOutputFilename = `${testPath.substring(
+        0,
+        testPath.lastIndexOf('/')
+    )}/__image_snapshots__/${currentTestName}-diff.png`;
+    const diffPercentage = (result * 100) / (width * height);
+    const pass = diffPercentage <= 0.05;
+
+    if (!pass && writeDiff) {
+        fs.writeFileSync(diffOutputFilename, PNG.sync.write(diff));
+    } else if (fs.existsSync(diffOutputFilename)) {
+        fs.unlinkSync(diffOutputFilename);
+    }
+
+    return { message: () => `Images were ${result} (${diffPercentage.toFixed(2)}%) pixels different`, pass };
+}
