+++ libraries/ag-charts-test/src/index.ts
@@ -1,4 +1,5 @@
 export * as mockCanvas from './canvas/mock-canvas';
+export * from './canvas/to-match-image';
 export * from './examples/load-example';
 export * from './console/mock-console';
 export * from './benchmarks/timing';
