+++ packages/ag-charts-community/package.json
@@ -84,8 +84,6 @@
     "ag-charts-test": "11.3.1-beta.20250611",
     "skia-canvas": "^2.0.2",
     "path2d": "^0.2.0",
-    "pixelmatch": "^5.2.1",
-    "pngjs": "3.4.0",
     "timezone-mock": "^1.3.6"
   },
   "publishConfig": {
