+++ packages/ag-charts-community/src/util/test/mockCanvas.ts
@@ -1,13 +1,8 @@
-import * as fs from 'fs';
-import pixelmatch from 'pixelmatch';
-import { PNG } from 'pngjs';
 import { Canvas } from 'skia-canvas';
 
 import { resetIds } from 'ag-charts-core';
-import { mockCanvas } from 'ag-charts-test';
+import { CANVAS_HEIGHT, CANVAS_WIDTH, mockCanvas } from 'ag-charts-test';
 
-export const CANVAS_WIDTH = 800;
-export const CANVAS_HEIGHT = 600;
 export const CANVAS_TO_BUFFER_DEFAULTS = { quality: 1 };
 
 export function extractImageData({
@@ -66,29 +61,4 @@ export function setupMockCanvas({ width = CANVAS_WIDTH, height = CANVAS_HEIGHT }
     });
 
     return mockCtx.ctx;
-}
-
-export function toMatchImage(this: any, actual: ImageData, expected: ImageData, { writeDiff = true } = {}) {
-    // Grab values from enclosing Jest scope.
-    const { testPath, currentTestName } = this;
-
-    const width = CANVAS_WIDTH;
-    const height = CANVAS_HEIGHT;
-    const diff = new PNG({ width, height });
-    const result = pixelmatch(actual.data, expected.data, diff.data, width, height, { threshold: 0.01 });
-
-    const diffOutputFilename = `${testPath.substring(
-        0,
-        testPath.lastIndexOf('/')
-    )}/__image_snapshots__/${currentTestName}-diff.png`;
-    const diffPercentage = (result * 100) / (width * height);
-    const pass = diffPercentage <= 0.05;
-
-    if (!pass && writeDiff) {
-        fs.writeFileSync(diffOutputFilename, PNG.sync.write(diff));
-    } else if (fs.existsSync(diffOutputFilename)) {
-        fs.unlinkSync(diffOutputFilename);
-    }
-
-    return { message: () => `Images were ${result} (${diffPercentage.toFixed(2)}%) pixels different`, pass };
 }
