+++ packages/ag-charts-enterprise/jest.setup.ts
@@ -4,8 +4,7 @@ import { DOMMatrix, Image, Path2D } from 'skia-canvas';
 import { URL } from 'url';
 import { TextDecoder, TextEncoder } from 'util';
 
-import { toMatchImage } from 'ag-charts-community-test';
-import { mockCanvas } from 'ag-charts-test';
+import { mockCanvas, toMatchImage } from 'ag-charts-test';
 
 // @ts-expect-error types don't exactly align
 global.Canvas = mockCanvas.ConfiguredCanvas;
